auth: handle SyntaxError edge case in refresh credentials
Problem:

There is a small edge case where there are missing fields in a SyntaxError,
in this case the $response or its reason field do not exist. Due to this those
cases are causing refresh credentials to fail.

Solution:

Handle the case where those fields are missing and gracefully
handle them by still marking it as an AwsClientResponseError.
Since we cannot extract the underlying error message, we set
a custom errors message.

Now this will not cause a failed refresh credential to invalidate.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>