auth: handle sso cache errors gracefully
Problem:

When we get SSO cache errors, such as a failed write, we do not want to
invalidate the connection. Instead we want to gracefully handle this similar
to network errors, since the connection can still be recovered on a subsequent
token refresh.

Solution:

Catch all sso cache errors and wrap them in a high level DiskCache error so
that we can easily determine an sso cache error during token refresh. Then we can
know not to invalidate the connection when there is a cache error.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

show error message on refresh file system error

This shows the user an error message when refresh token fails,
it will have a button to open the logs which hopefully they can use to debug the issue.

There is also a button to never show the message again in the case for users who may not
want to immediately fix the issue, and we do not want to bother them.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>