fix(amazonq): sign-in notification displaying if user is signed in
Problem: `Start using Amazon Q` can display if there is some error restoring the connection object in the main Auth class. May be a network error or some other error. This is fine because Q uses secondary auth, so it only needs to be restored in that class. However, the notificaiton logic access from the Auth class which may not be restored.

Solution: Use AuthUtil to check if there is a connection instead of Auth.