feat(amazonq): Reduce @workspace indexing time by 50%  (#5519)
## Problem

The indexing of @workspace feature is slow.

## Solution

By upgrading ONNX version from 1.18 to 1.19 in the 0.1.5 LSP, we
observed reduction in indexing time by about 50% when using CPU for
index. There is no improvement when using GPU or AMD CPU on Windows.
There is no customer facing change or customer experience change.



---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.