feat(lambda): [Getting Started]Setting stackNode to display current stack (#1676)
## Problem

To display the current stack that user used on the sidebar. 

## Solution

We will store the user's currently used `stackName` and `region`,
whether it is for "deploy" or "sync" in the `samconfig.toml` file's
`default.global.parameters` and will show them on the sidebar.
The global parameters will get updated every time user use deploy or
sync functionality.

For the situation when there is no global parameters in samconfig file,
the node will be removed.