fix(auth): filesystem error during SSO token refresh expires session #5549
## Problem:
When we get SSO cache errors, such as a failed write, we do not want to
invalidate the connection. Instead we want to gracefully handle this
similar
to network errors, since the connection can still be recovered on a
subsequent
token refresh.

A place that this happens is during SSO token refresh. Where if there is
a filesystem
error when writing the new token to disk, the connection will become
invalid.

## Solution:
Catch sso cache errors and wrap them in a high level `DiskCacheError` so
that we can easily determine an sso cache error during token refresh.
Then we can
know not to invalidate the connection when there is a cache error since
`isRecoverableError()` will
catch the error.

Additionally a message is shown to the user explaining how the sso token
refresh failed and guides
them to the logs in hopes that they can fix it themselves, or at least
give it to us when the report the error.