refactor(telemetry): prepend the error code/name to reasonDesc field
Problem:

In telemetry we sometimes add the error message to the `reasonDesc` field.
Additionally when using `getTelemetryReasonDesc()` it will recurse through the causal
chain of the error and return a value composed of all the error messages.

The issue is that the nested error messages may have useful context in their
id (code or name) that is lost.

Solution:

Prepend the code/name to the error messages when using `getTelemetryReasonDesc()`

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>