refactor(auth): redo DiskCacheError
Problem:

The original implementation of DiskCacheError can be simplified to achieve
the original purpose of what we want.

Solution:

In the root of where the filesystem errors occured during token refresh we
will wrap those errors in a DiskCacheError. This will allow anything upstream
to know if they specifically got a DiskCacheError, which in the context of refreshing
a token means that a cache operation failed, implying that the SSO session is not
actually stale.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>