fix(getting started): Fix Appbuilder failure related to invalid value in the Config (#1709)
## Problem
I wanted to add the region to my samconfig, and I mistakenly didn’t add
quotes to the region name. That caused that when I refreshed the app
builder, I got an error message saying “Unable to load Resource tree for
this App. Update IaC template”. Which is not correct, since the problem
was in my samconfig and not on the IaC template

## Solution
fix how the app builder handles exception