fix failing test
due to this change one of the tests needed a slight update to work again due
to a race condition that occurs during testing since we do not know when all
the required event emitters are fully completed.

In this case the onDidDeleteConnection event handler in SecondaryAuth was async
and the unit test completed before the handler finished. This caused the test to fail.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>