fix(lambda): [Getting Started]Correcting fields for Local and Remote invoke (#1712)
## Problem
This PR addresses the handling of parameters required for invoking
remote test events. To invoke a remote test-events it needs either:
1. `FunctionArn`, or
2. `StackName` and `LogicalId`.

Different entry points provide different sets of parameters:
1. From the "Application Builder" explorer, both `StackName` and
`LogicalId` are available.
2. From the "AWS" explorer, `FunctionArn` is available.

Since both entry points utilize the same function, we need to ensure
that all three parameters (StackName, LogicalId, and FunctionArn) are
passed to the function to accommodate both entry points effectively.

## Solution
This PR adds all three properties to pass to the main function and all
other respective places that leads to the main function.