ci: remove src.gen from workspaces to fix dependabot
Problem: I theorize that we are still getting dependabut for src.gen because it is a subworkspace of the root project.

Solution: Remove it and install/build separately on root `npm i`. Use silent so that we don't fail for deprecation errors in its dependencies (we don't manage these deps).