fix(amazonq): Use the correct state for Q in remote environment (#5575)
## Problem:

As part of our design we intentionally separate the Auth state between a
local environment, versus a remote (ssh'd) one. We do this so that if
the user has 2 IDE instances, one local and the other remote, if one
changes its auth state the other one will not be affected.

The reason for separating auth that I can remember is due to how we
store the actual tokens on disk. Because the remote does not have access
to the disk of the local, we need to ensure the remote operates its auth
independently of the local auth.

By design globalState is shared by both local and remote instances. So
for local we need to ensure we do not use the base globalState,
otherwise the same Auth state as the remote will be used.

The problem, SecondaryAuth is not doing this for all cases. It is using
the globalState when it should be asking Auth which state it should use.
The happy path was working though, but there can be potential issues
with this bug at some other point.

## Solution:

Expose the state that Auth is using so that something like SecondaryAuth
can ask Auth for the correct state object depending on the local vs
remote.


---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>
Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>