fix(lambda): [Getting Started] Show File name in webview instead of content (#1729)
## Problem
Local invoke - When specifying a file, the textarea is still editable but it should be read-only. 

## Solution
The intention is that the file is modified elsewhere because the editing experience is more rich. This also helps avoid accidential event modification if customers think the payload is stored in their file, but currently we're just copying the contents into the textarea.