fix(auth): wait to display Q sign-in notification
(Super low risk and effort fix, but may not address the root issue due to difficulties to repro)

Problem: The sign-in notification is queued to display immediately after startup. Some users are reporting that it displays even if they are signed in. This may be due to the tryAutoConnection routine not having a chance to complete because of its async nature.

Solution: Put a delay on displaying the notification to users to potentially mitigate this issue.