feat(ec2): prompt to add inline policies.  (#3705)
## Problem
By default, the ec2 instances do not have the proper permissions to
allow for connect via SSM. Therefore, users will have to complete this
step manually.
## Current Solution
Check if IAM role exists on instance, if it does, prompt to add
necessary actions as inline policy.
### Details and Problems with Current Solution
- actions in prompt are poorly formatted. 
- attaching the inline policy, then immediately attempting connection
does not appear to work reliably. The cause of this is currently
unknown, but one potential explanation is the delay of propagating
permissions across all regions. When immediately attempting connection
after adding the inline policy, the terminal will open, but will
immediately crash. Note that no error is thrown by the SDK here, meaning
we have no way of knowing if the terminal will crash or not before
opening the terminal. Additionally, while we do get a terminal close
event on the crash, we have no way of distinguishing this from a user
close event. The current solution to this problem is to add a 5 second
delay between adding the policies and attempting connection. During this
process we provide a progress bar with cancel. This solution is far from
ideal, but does appear to work. A polling approach was also attempted,
but seemed to fail as the SDK calls returned that the necessary policies
were attached before a connection can be successfully made.
- after attaching the inline policy, there is often a delay in the SSM
agent starting on the instance. This results in receiving the `is SSM
agent running error?` if connection is attempted immediately following
adding the inline-policy (only in remote window, this error doesn't
happen when trying to establish a terminal). This can take up to a few
minutes to go away, or can be immediately resolved by fully restarting
(not rebooting) the instance. Giving the customer an option to restart
is not a feasible solution as it comes with many side effects (new
billing cycle, new ip address, etc.).
- Actions added include 
```
    'ssmmessages:CreateControlChannel',
    'ssmmessages:CreateDataChannel',
    'ssmmessages:OpenControlChannel',
    'ssmmessages:OpenDataChannel',
    'ssm:DescribeAssociation',
    'ssm:ListAssociations',
    'ssm:UpdateInstanceInformation'
```
- The bottom three, while not necessary for the connection, appear to
help in starting the SSM agent on initial connection. With these
policies attached, and restarting the instance, the connection works
every time.
- Actions added are expanded such that SSM agent starts on instance. 
- If no IAM role attached, we link to documentation. 

### Visuals

progress bar with cancel:
<img width="539" alt="image"
src="https://github.com/aws/aws-toolkit-vscode/assets/42325418/b144ed2e-5bf2-41a2-b7f2-0f56e598ea10">

prompt with actions: 
<img width="306" alt="image"
src="https://github.com/user-attachments/assets/c9c0aa1c-8e13-4c4d-a178-f503e633647d">


<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Update the changelog using `npm run newChange`.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: JadenSimon <31319484+JadenSimon@users.noreply.github.com>
Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>
Co-authored-by: Weinstock <hkobew@80a9971f0a95.ant.amazon.com>