determine if in Debug & Run mode
When we run our extension in Run & Debug mode (clicking the green play button)
it spins up a new IDE instance and runs our extension there.

The reason we need to know this is that for crash reporting we do not want
to emit telemetry in a debug window due to it not gracefully shutting down
if the developer clicks the red stop button.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>