build: detach src.gen/ from monorepo
Theory: dependabot is still trying to update src.gen/ dependencies because they are managed by the root monorepo. So technically, there is no way to avoid updating src.gen's package.json because the root package.json is linked to it and impacted.

Let's make its own "subrepo" and see if we can get dependabot to ignore it finally.