build: detach src.gen/ from monorepo (#5667)
Theory: dependabot is still trying to update src.gen/ dependencies
because they are managed by the root monorepo. So technically, there is
no way to avoid updating src.gen's package.json because the root
package.json is linked to it and impacted by dependabot changes for
subproject dependencies.

Let's make its own "subrepo" and see if we can get dependabot to ignore
it finally.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.