test(ec2): fix flaky tests regarding child node updates.  (#5698)
## Problem
The issue is here:
https://github.com/aws/aws-toolkit-vscode/blob/9eea1f3700930df79620ef26ff6a43450e593580/packages/core/src/awsService/ec2/explorer/ec2ParentNode.ts#L55-L64
This fires off multiple asynchronous calls on each child, with the
function returning before any are resolved. Thus, the child node could
be deleted from the set after this function returned (potentially when
another piece of code is using it). This leads to `undefined` errors
within that function.
## Solution
Refactor the code to make use of a `for...of` with `await` pattern. 

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.