fix(lambdd): [Getting Started] error handling, updating samconfig.toml
## Problem
Auto refresh feature trigger extension to fetch for deployed stack node
using `stack_name` and `region` from `samconfig.toml` file. In current
logic, the `stack_name` and `region` are written before deployment
process completed to avoid the edge case when deployment failure due to
stack up-to-date. However, this order triggers the extension to fetch
for stack that may not already been created and through an error toast
message to customer stating "Stack does not exist".

## Solution
- Add additional logic to check for deployment error message to see if
deployment failure is due to stack is up-to-date
- Add try-catch block to still update the samconfig.toml if deployment
failure is due to stack is up-to-date
- Move the process of writing to `samconfig.toml` after deploy/sync
process completes/fails
- Remove toast message for stack does not exist case
- Add logic for handling empty array when parsing `StackResource[]` to
avoid displaying error message.