refactor(urls): Remove isCloud9 from constants file
The isCloud9() function was causing issues in the constants file.
The constants file should be a minimal module with as little external
dependencies as possible.

We had some strings that needed to resolve a url depending on c9 vs toolkit.

Solution:

Move the resolution of c9 vs toolkit docs to an external function. Update uses
everywhere.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>