refactor(urls): Remove isCloud9 from constants file (#5717)
We have some strings that need to resolve a url depending on c9 vs
toolkit.

The isCloud9() function is causing issues in the constants file. Due to
some sort of module import order, isCloud9() did not resolve
appropriately when expected. I think we used the constants file before
all the dependencies of the isCloud9() function were available

The constants file should be a minimal module with as little external
dependencies as possible.

## Solution:

Move the resolution of c9 vs toolkit docs to an external function.
Update uses everywhere.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>