fix(crash): crash monitoring dir not being cleaned up
During development the underlying implementation was changed to get
the OS uptime. The initial implementation returned uptime in minutes,
but the new one returned it in seconds and this was not accounted for.

As a result the crash monitoring folder was not being cleaned up when expected
on computer restart

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>