feat(lambda): [Getting Started] Display all remote resources, actions
We wanted to display the context menu that appears for each node in the
Explorer section in the new Appbuilder section to have a consistent user
experience. However, since the Explorer section nodes have TreeItem as
their base class and we are using TreeNode as the base class for all
nodes in the Appbuilder section, we needed to transform our TreeNodes
into TreeItems in all activation files before we can call the existing
command functions. Presently the same context menus are implemented for
functions, s3 buckets and APIs but we will add support for all the nodes
supported in the Explorer in the coming weeks.