fix(amazonq dev): user can view diffs of previous /dev iterations #5756
## Problem
We currently only support accessing the last code generated (by the Tree
view). If you try open the diff view in past history, will thrown an
error. This happens because when we click on past history Tree we
actually send a "newer" uploadId and didn't keep track of all the past
ones.

## Solution
Cache those uploadIds based on codeGenerationId which can be easily
accessible while clicking in past views.