telemetry(amazonq dev): track empty path file scenario #5625
## Problem
When customer use workspaces to generate files, RTS and LLM cannot
decide which folder to deliver the generated file. We're for now
tracking how many times we reach for accountability and prioritization.

## Solution
Add telemetry to count and save the conversationIds that reach this
specific workspace usage.

PS: This is only included in VS Code since the Workspaces capability is only used here.