feat(notifications): initial rule engine + tests
Initial code for a rule engine to determine whether or not to show an in-IDE notification.
A notification is a JSON payload with a set amount of criteria. The rule engine accepts
context from the currently running extension then determines if the notification payload's
criteria will fit the provided context.

The types match the commonly designed schema, but may change in future commits.

Future work:
- More docs
- Updates to types and/or criteria
- Code that will use this