test(amazonq): implement performance test for downloadExportResultArchieve.  (#5710)
## Problem
Continuation of work here:
https://github.com/aws/aws-toolkit-vscode/pull/5670

## Solution
In this case `downloadExportResultArchieve` is reading a large
collection of objects into a buffer, then writing that buffer to a file.
We test this with varying collections of objects. We test 1x1KB,
10x100B, 100x10B, and 1000x1B.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.