fix(crashMonitoring): handle sleep/wake appropriately
Problem:

When a user does a sleep then wake, our heartbeats are not up to date since
they cannot send when the user's computer is asleep.

Due to this there is a race condition between the heartbeats sent and when we check the
heartbeats to determine if they are stale (crash).

Solution:

Keep a TimeLag class that helps to assert that there is no lag in our time.
It works by updating a state every second. And if we determine that the next
update to that state took longer than a second, we determine that there was a lag.
Then we simply skip the next crash check, allowing a fresh heartbeat to be sent.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>