fix(test): flaky auth test
This test was flaky due to the event emitter and the test expecting 2
of the same event to come in, but sometimes the second one was not capture.

This looks to be due to the second `captureEventOnce` call needing to do some
initial setup before it could capture an event. And I think there was a
race condition between it being setup in time and the event being emitted before
it could start listening.

Solution:

Make a new function which does the event capture once and then listens N
amount of times. This is reliable.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>