fix: handle edge case of irrelevant files
Problem:

Previous versions of the crash checker added files to the folder which we use
to maintain state. Some of those files are now irrelevant.

The issue is that we did a naive readdir of the folder and incorrectly used these files

Solution:

Only read the relevant files and ignore the others

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>