fix(telemetry): session_end not caching events
When the user shutdown their extension gracefully our mechanism that
would cache any pending metrics for the next startup would not save to disk.

This was due to the VS Code FS implementation being used, but it will not work since
the function it is called in is called by `deactivate()` which does not have access to
vscode apis on shutdown.

Solution is to swap it with the node fs implementation

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>