ci: don't output code coverage table in logs (#5813)
Problem: c8 was not given a reporter, therefore it defaulted to `text`
which is the terminal.
Solution: Set it to `lcov`, which is what we have been using previously.

Note: More work may be needed on the .c8rc.json files. We might not need
the one for core/ anymore. They may not be configured in the optimal
way. Is it behaving how we want it to?

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.