fix(packageJson): Sync fonts to Toolkit
Problem:

A new font was added for Amazon Q by adding the SVG to `packages/core/resources/icons`.
Then the `generateIcons.ts` script will read this and add the icon to our custom font that we generate.
Another part of this script also updates the package.json with this new icon.

The problem is that the package.json change was only reflected in the Amazon Q package.json
and not the Toolkit package.json. This needs to be reflected in the Toolkit package.json as well
since we only generate a single font for both extensions, instead of a customized font for each.

So every time I did a development build the package.json would change.

Solution:

Run the generate icon script for Toolkit and commit the package.json changes.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>